+++ src/models/dto/login.dto.ts
@@ -1,11 +1,5 @@
-import { IsEmail, IsString, Length } from 'class-validator';
-
 class LoginDTO {
-    @IsEmail()
     username: string;
-
-    @IsString()
-    @Length(8)
     password: string;
 }
