+++ src/routes/index.ts
@@ -1,1 +1,1 @@
-export * from './auth.routes';
+export * from './auth/auth.routes';
